+++ modules/tinymce/src/core/test/ts/browser/content/EditorContentTest.ts
@@ -241,6 +241,13 @@ describe('browser.tinymce.core.content.EditorContentTest', () => {
           assertEventsContentType();
         });
 
+        it('TINY-9140: Insert content without firing events', () => {
+          const editor = hook.editor();
+          editor.setContent('<p>html</p>', { no_events: true });
+          assertEventsFiredInOrder([]);
+          assertEventsContentType();
+        });
+
         it('TINY-7996: Set tree content with content altered in BeforeSetContent', () => {
           const editor = hook.editor();
           editor.setContent('<p>tree</p>');